improved modal, various bug fixes
